more cleanups

+++ src/utils/timestamp.c
@@ -4,7 +4,7 @@
 static unsigned int dotm365[13] = {0, 31, 59, 90, 120, 151, 181, 212, 243, 273, 304, 334, 365};
 static unsigned int dotm366[13] = {0, 31, 60, 91, 121, 152, 182, 213, 244, 274, 305, 335, 365};
 
-static unsigned int isleapyear(unsigned short year)
+static unsigned int isleapyear(unsigned int year)
 {
 
     if ((year % 4) != 0)
